Use new test syntax for Avatar component
